Define ellmer tools from MCP servers

These functions implement R as an MCP client, so that ellmer chats can register functionality from third-party MCP servers such as those listed here: https://github.com/modelcontextprotocol/servers.

mcp_tools() fetches tools from MCP servers configured in the acquaint server config file and converts them to a list of tools compatible with the $set_tools() method of ellmer::Chat objects.

Usage

mcp_tools(path = NULL)

Arguments

path

A single string indicating the path to the acquaint MCP servers configuration file. If one is not supplied, acquaint will look for one at the file path configured with the option .acquaint_config, falling back to file.path("~", ".config", "acquaint", "config.json").

Value

mcp_tools() returns a list of ellmer tools that can be passed directly to the $set_tools() method of an ellmer::Chat object. If the file at path doesn't exist, an error.

Configuration

acquaint uses the same .json configuration file format as Claude Desktop; most MCP servers will define example .json to configure the server with Claude Desktop in their README files. By default, acquaint will look to file.path("~", ".config", "acquaint", "config.json"); you can edit that file with file.edit(file.path("~", ".config", "acquaint", "config.json")).

The acquaint config file should be valid .json with an entry mcpServers. That entry should contain named elements, each with at least a command and args entry.

For example, to configure mcp_tools() with GitHub's official MCP Server https://github.com/github/github-mcp-server, you could write the following in that file:

{
  "mcpServers": {
    "github": {
      "command": "docker",
      "args": [
        "run",
        "-i",
        "--rm",
        "-e",
        "GITHUB_PERSONAL_ACCESS_TOKEN",
        "ghcr.io/github/github-mcp-server"
      ],
      "env": {
        "GITHUB_PERSONAL_ACCESS_TOKEN": "<add_your_github_pat_here>"
      }
    }
  }
}

See also