Model context protocol for your R session

Together, these functions implement a model context protocol server for your R session.

Usage

mcp_host()

mcp_server()

Configuration

mcp_server() should be configured with the MCP clients via the Rscript command. For example, to use with Claude Desktop, paste the following in your Claude Desktop configuration (on macOS, at file.edit("~/Library/Application Support/Claude/claude_desktop_config.json")):

{
  "mcpServers": {
    "r-acquaint": {
      "command": "Rscript",
      "args": ["-e", "acquaint::mcp_server()"]
    }
  }
}

Or, to use with Claude Code, you might type in a terminal:

claude mcp add -s "user" r-acquaint Rscript -e "acquaint::mcp_server()"

mcp_server() is not intended for interactive use.

The server interfaces with the MCP client on behalf of the host in your R session. Use mcp_host() to start the host in your R session. Place a call to acquaint::mcp_host() in your .Rprofile, perhaps with usethis::edit_r_profile(), to start a host for every interactive R session you start.

Examples

if (interactive()) {
mcp_host()
}